+++ src/Modele/Joueurs/Joueur.java
@@ -1,5 +1,8 @@
 package Modele.Joueurs;
 
+import java.util.ArrayList;
+
+import Modele.Plateau.Pingouin;
 import Modele.Plateau.Plateau;
 import Utils.Couple;
 import Utils.Position;
@@ -9,11 +12,13 @@ public abstract class Joueur {
 	private int nbPingouins;
 	private int scoreFish;
 	private int scoreDestroyed;
+	private ArrayList<Pingouin> pingouins;
 	
 	public Joueur(int id){
 		this.id = id;
 		this.scoreFish = 0;
 		this.scoreDestroyed = 0;
+		this.pingouins = new ArrayList<Pingouin>();
 	}
 	
 	public Joueur(int id,int p){
@@ -21,6 +26,7 @@ public abstract class Joueur {
 		this.nbPingouins = p;
 		this.scoreFish = 0;
 		this.scoreDestroyed = 0;
+		this.pingouins = new ArrayList<Pingouin>();
 	}
 
 	public int id() {
@@ -79,6 +85,14 @@ public abstract class Joueur {
 		this.nbPingouins -= l;
 	}
 
+	public ArrayList<Pingouin> pingouins(){
+		return this.pingouins;
+	}
+	
+	public void addPingouins(Pingouin p) {
+		this.pingouins.add(p);
+	}
+	
 	/**
 	 * Fonction appelee dans l'IA pour calculer le prochain coup
 	 * renvoie (PosPingouin,PosObjectif) si une erreur c'est produite
@@ -101,7 +115,14 @@ public abstract class Joueur {
 	 * @return
 	 */
 	public boolean posePingouin(Plateau plateau,Position position) {
-		return false;
+		boolean res;
+		Pingouin p = new Pingouin(this.id());
+		res = plateau.poserPingouin(position, p);
+		if(res) {
+			this.addPingouins(p);
+			this.addScoreFish(1);
+		}
+		return res;
 	}
 
 	/**
@@ -115,8 +136,25 @@ public abstract class Joueur {
 	 * @return le nombre de poissons mangee par le pingouin lors du coup ou -1 si une erreur c'est produite
 	 * @throws Exception 
 	 */
-	public int jouerCoup(Plateau plateau, Position start, Position goal) throws Exception {
-		return -1;
+	public int jouerCoup(Plateau plateau,Position start, Position goal) throws Exception {
+		int res;
+		
+		if(plateau.getCellule(start).aPingouin()) { //test si le pingouin existe
+			if(plateau.getCellule(start).pingouin().employeur() == this.id()) { //test si le pingouin appartient bien a ce joueur
+				res = plateau.jouer(start,goal);
+				if(res > 0) {
+					this.addScoreFish(res);	
+				}
+				if(res >= 0) {
+					this.addScoreDestroyed(1);	
+				}
+				return res;
+			} else {
+				throw new Exception("Le pingouin en "+start+" n'appartient pas au joueur "+this.id()+".");
+			}
+		} else {
+			throw new Exception("La case en "+start+" ne contient pas de pingouin.");
+		}
 	}
 	
 	public boolean estIA() {
